Various refactorings
Seperated apple sent but not rejected queue logic into a class call
PendingItemsQueue.
Created a class for collected the state of SentNotifications from a
PendingItemsQueue.
Corrected some issues with DeviceToken.
Moved SentNotification to it's own file.
Corrected bug in in ChannelScaler
Added a test to ensure that if a distributor with 1 open channel that is
coping with the load.  That the distributor is not scaled.